Remove unused code and update snapshots.
