+++ src/components/index.js
@@ -184,9 +184,6 @@ export {
 
 export {
   EuiSideNav,
-  EuiSideNavGroup,
-  EuiSideNavItem,
-  EuiSideNavTitle,
 } from './side_nav';
 
 export {
