+++ src/components/side_nav/_index.scss
@@ -1,4 +1,2 @@
 @import 'side_nav';
 @import 'side_nav_item';
-@import 'side_nav_title';
-@import 'side_nav_group';
