--- src/components/side_nav/_side_nav_group.scss
@@ -1,27 +1,0 @@
-// .euiSideNavGroup {
-//   position: relative;
-//   margin-left: $euiSize;
-
-//   &:after {
-//     position: absolute;
-//     content: "";
-//     top: 0;
-//     bottom: $euiSizeM;
-//     width: 1px;
-//     background: $euiBorderColor;
-//     left: 0px;
-//   }
-
-//   .euiSideNavItem {
-
-//     &:after {
-//       position: absolute;
-//       content: "";
-//       top: 50%;
-//       left: 0;
-//       width: $euiSizeXS;
-//       height: 1px;
-//       background: $euiBorderColor;
-//     }
-//   }
-// }
