+++ src/components/side_nav/_side_nav_item.scss
@@ -130,24 +130,3 @@
     background: $euiBorderColor;
   }
 }
-
-// .euiSideNavItem {
-//   @include euiFontSizeS;
-
-//   display: block;
-//   position: relative;
-//   width: 100%;
-//   text-align: left;
-//   color: $euiColorDarkShade;
-//   white-space: nowrap;
-//   overflow: hidden;
-//   text-overflow: ellipsis;
-//   font-size: $euiFontSizeS;
-//   padding: $euiSizeXS / 2 $euiSize $euiSizeXS / 2 $euiSizeS;
-
-//   &.euiSideNavItem--parent {
-//     border-right: solid $euiSizeXS transparent;
-//   }
-
-
-// }
