+++ src/components/side_nav/index.js
@@ -1,4 +1,1 @@
 export { EuiSideNav } from './side_nav';
-export { EuiSideNavGroup } from './side_nav_group';
-export { EuiSideNavItem } from './side_nav_item';
-export { EuiSideNavTitle } from './side_nav_title';
